Move all function related to PythonJob to the PythonJob Task
